Move slash commands to a seperate file
